Add tasty, hspec, hspec-disover, and add one test that fails
This test should probably fail when it throws because of a missing Anki
database. This test can be fixed by replacing the default with something
that ships with the test suite. The Default instance should probably be
moved to the test suite, since a library cannot easily assume where on a
system something is installed.

```
test/test.hs
  getUserProfile
    getUserProfile
      somehow handles Nothing: FAIL
        uncaught exception: SQLError
        SQLite3 returned ErrorCan'tOpen while attempting to perform open
          "/home/simon/Documents/Anki/prefs.db": unable to open database file
```